Have shortcuts use static app.ico instead of that of the ephemeral exe
using the app.ico in the root folder fixes the problem of losing the shortcut icon when we upgrade, lose the original, and eventually the windows explorer cache loses it. There was another attempt at fixing this by finding all the shortcuts and updating them, but that didn't work in our testing and this seems simpler and more robust. There may be some other reason for the old approach of pointing at the icon of the app itself (e.g. could be a different icon)?