Get shortcut icon from <ExeName>.ico
The previous version used app.ico as the main application icon.
This forces the icon used for the program to be the one used
for setup and uninstall. Also, it may be missing, since it is
obtained from the internet. This version expects the current
install to have an icon with the same path as the EXE
(e.g., BloomAlpha.ico) and copies it to the root folder so
we don't have to update the shortcut for a new version (though
we do update the icon, since it may have been changed).