Repair --rerunningWithoutUAC
When Squirrel is run in admin mode (without --allUsers), it re-runs
itself with this additional command line argument (and no admin
privileges, which can cause problems for standard install).
Unfortunately the extra argument is passed on to Update.exe
which does most of the work, and it rejects it as unknown.
With this change it is just ignored.